servo_ticks_per_microseconds calc should be float

+++ src/JMotorDriver/JMotorDriverEsp32Servo.h
@@ -52,7 +52,7 @@ public:
         }
         SERVO_FREQ = freq;
         SERVO_RES = resBits;
-        SERVO_TICKS_PER_MICROSECOND = (1 << SERVO_RES) * SERVO_FREQ / 1000000; //DEFAULT=52.4288  2^SERVO_RES / 1E6 * SERVO_FREQ
+        SERVO_TICKS_PER_MICROSECOND = (1 << SERVO_RES) * (float)SERVO_FREQ / 1000000; //DEFAULT=52.4288  2^SERVO_RES / 1E6 * SERVO_FREQ
         unsigned long startMicros = micros();
         while (digitalRead(servoPin) == HIGH && micros() - startMicros <= maxServoValue)
             ; //wait for pulse to go low to avoid cutting it short and causing the servo to twitch
